+++ computers/src/main/java/dev/jstech/computers/machine/MachineServices.java
@@ -257,9 +257,9 @@ public final class MachineServices implements IHost {
             this.shell = new ServerCliComputer(terminal, server);
             this.files = new FileService(this.shell);
             this.computer = new ComputerInfoService(this.machine, this.shell);
-            this.network = new NetworkReadService(terminal, server, this.shell);
-            this.mainframe = new MainframeStatsService(this.shell);
-            this.operations = new OperationsService(this.shell);
+            this.network = new NetworkReadService(terminal, server, this.shell, this.shell);
+            this.mainframe = new MainframeStatsService(this.shell, this.shell);
+            this.operations = new OperationsService(this.shell, this.shell);
             this.iql = new IqlService(this.shell);
             this.programs = new ProgramService(this.machine, terminal, server, this.shell);
             this.remotes = new RemoteComputerService(this.shell);